// sisCT/src/app/interfaces_crud/recursos-humanos.interface.ts
export interface RecursoHumano {
    pkRecursosHumanos: number;
    dsNome: string;
    nuCpf: number;
    dtNascimento: Date;
    dsTelefone: string;
    dsEmail: string;
    dsCargo: string;
    tpFormacaoAcademica: string;
    nuCargaHoraria: number;
    tpVinculo: string;
    stAtivo: number;
    dtUltimaAtualizacao: Date;
    pkCadastroNacional: number;
}